requirements.txt
bcrypt
blinker
cffi
Click
Flask
Flask-Bcrypt
Flask-Login
Flask-Mail
Flask-SQLAlchemy
Flask-WTF
itsdangerous
Jinja2
MarkupSafe
Pillow
pycparser
six
SQLAlchemy
Werkzeug
WTForms